pyproject.toml
[project]
name = "robot-navigation"
version = "0.1.0"
description = "Raspberry Pi robot with motor and camera"
authors = [
    {name = "rayane.mechik",email = "rayane.mechik@viziosense.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.9"
dependencies = [
    "colorzero==2.0",
    "gpiozero==2.0.1",
    "numpy==1.24.2",
    "opencv-python==4.11.0.86",
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "rpi-gpio (>=0.7.1,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
